refactor: split uml.rb into individual files
Split 508-line uml.rb into 18 one-class-per-file modules under
lib/xmi/uml/ with autoload. Pure file reorganization, zero logic
change.
